refactoring: compiler warnings has been fixed
for pod namespace and for gpstk classes specific for this branch

+++ ext/lib/Geomatics/ARLambda.cpp
@@ -43,7 +43,7 @@
           if( lambda(ambFloat,ambCov,F,S,2)==0 )
           {
              Vector<double> ambFixed( ambFloat.size(), 0.0 );
-             for(int i=0; i<ambFloat.size(); i++) 
+             for(size_t i=0; i<ambFloat.size(); i++) 
              {
                 ambFixed(i) = F(i,0);
              }
